Start on a new, sparse-set based, sharded ECS for Feather.
The intention behind this new project is to allow for sharing components
across threads and, in the future, allowing Feather to perform region-based
sharding.

This is a long-term project. I intend to work on our new ECS gradually
over the next few months. It will not be incorporated into Feather
until then.